+++ common/debug.cpp
@@ -32,7 +32,7 @@ using Atlas::Message::Element;
 using Atlas::Message::MapType;
 using Atlas::Message::ListType;
 
-void output_element(std::ostream & out, const Element & item, int depth)
+void output_element(std::ostream & out, const Element & item, size_t depth)
 {
     switch (item.getType()) {
         case Element::TYPE_INT:
